Make sure plugin-based installations keep working. Fixes #52
`gemspec by default uses the path . when looking for the gemspec. That's fine
`when it's being bundled on its own, but when bundling via Redmines Gemfile, "."
`refers to the path to the Redmine installation.

Without this change, Redmine installations having the plugin installed directly
in plugins/redmine_github_hook would get an error like

> There are no gemspecs at /Users/jakob/Projects/redmine/redmine-2.5.

when starting Redmine (or when starting Bundler, really)